+++ src/main/java/com/example/dalyda_backend_stockmanager/services/SalesService.java
@@ -2,12 +2,13 @@ package com.example.dalyda_backend_stockmanager.services;
 
 import com.example.dalyda_backend_stockmanager.dtos.SalesDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface SalesService {
     Integer getWeeklySales();
 
-    Integer getWeeklySalesValue();
+    BigDecimal getWeeklySalesValue();
 
     List<SalesDto.RecentSaleResponse> getRecentSales();
 }
